+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to Family Diagram are documented in this file.
 
 ## 2.1.17b1
 
+- Added Triangle view, click the triangle badge on a mover to show positions
 - Fixed broked "Add birth" and "Edit birth" buttons in person properties
 - Fixed person picker in add form not showing renamed people in auto-complete.
 - Can now use down/up arrows + enter/return to pick a person via autocomplete in add event form.